refactored close into guard
