--- src/library/lib_guard.rs
@@ -1,2 +1,0 @@
-use super::*;
-impl<L: Loader> LibraryGuard<'_, L> {}
